requirements.txt
tensorflow==2.19.0
streamlit
numpy
Pillow
